requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# OpenAI Integration
langchain-openai
openai
langchain_huggingface
faiss-cpu


# Azure
azure-identity


#VectorDbs
langchain-pinecone

#loaders
pypdf
unstructured
azure-storage-blob
unstructured[docx]
langchain_chroma
langchain_experimental


## Tool Calling
wikipedia
youtube_search
langchain_tavily 
duckduckgo-search
pyowm


## LangGraph
langgraph

yfinance

libmagic


amadeus


streamlit
matplotlib